+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -12,6 +12,7 @@ group: Deployment & Configs
 
 ### RC
 
+- Fixed 'Host Boundary Profile' when used on walls that already have an Sketch.
 - Fixed inconsistency in 'Add Wall (Profile)' curve handling.
   [#1355](https://github.com/mcneel/rhino.inside-revit/issues/1355)